Extract HLL related code to separate module
