+++ hll/src/main/java/io/druid/hll/HLLCV1.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package io.druid.query.aggregation.hyperloglog;
+package io.druid.hll;
 
 import java.nio.ByteBuffer;
