+++ hll/src/test/java/io/druid/hll/HyperLogLogSerdeBenchmarkTest.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package io.druid.query.aggregation.hyperloglog;
+package io.druid.hll;
 
 import com.carrotsearch.junitbenchmarks.AbstractBenchmark;
 import com.carrotsearch.junitbenchmarks.BenchmarkOptions;
