+++ processing/pom.xml
@@ -37,6 +37,11 @@
         </dependency>
         <dependency>
             <groupId>io.druid</groupId>
+            <artifactId>druid-hll</artifactId>
+            <version>${project.parent.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>io.druid</groupId>
             <artifactId>bytebuffer-collections</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
