+++ processing/src/main/java/io/druid/query/aggregation/hyperloglog/HyperUniquesAggregator.java
@@ -19,6 +19,7 @@
 
 package io.druid.query.aggregation.hyperloglog;
 
+import io.druid.hll.HyperLogLogCollector;
 import io.druid.query.aggregation.Aggregator;
 import io.druid.segment.ObjectColumnSelector;
 
@@ -55,7 +56,7 @@ public class HyperUniquesAggregator implements Aggregator
   public Object get()
   {
     // Workaround for OnheapIncrementalIndex's penchant for calling "aggregate" and "get" simultaneously.
-    return HyperLogLogCollector.makeCollector(collector.getStorageBuffer().duplicate());
+    return HyperLogLogCollector.makeCollectorSharingStorage(collector);
   }
 
   @Override
